Lowercasing package names
The INI processing was updated and now the keys are stored in uppercase
instead of in a mixed case. This was for deduplication but it also
affected this program. Since INI file keys are supposed to be
case-insensitive, the rule is now made that all package names are in
lowercase.